create or use a file in memory and apply patch
Add a command argument --patch that can be used to update a file in
memory without saving it.
the command begin with fullpathfilename follow by a list of action (each
action begin with @ and must be terminated by the byte value 01.
3 actions exists:
- patch delete : @-pos_in_byte:byte_len_to_delete\001
if byte_len==-1 the buffer is erased till the end

- patch insert : @+pos_in_byte:string_code_to_insert\001

- remove file in memory : @x\001

at the first call if the file you want to patch exist on the disk it
will be read an copied into the buffer

ex:.
--patch i:/my/path/Foo.hx@-4:10\001@+15:my beautifull code i want to
insert